Moved routes into a separate file [Jerry Chang]
